pyproject.toml
[tool.poetry]
name = "aaron_common_libs"
version = "0.3.0"
description = "A collection of Python classes commonly used in my scripts."
authors = ["Aaron Melton <aaron@aaronmelton.com>"]
readme = "README.md"
homepage = "https://github.com/aaronmelton/aaron-common-libs"
repository = "https://github.com/aaronmelton/aaron-common-libs"
packages = [{include = "aaron_common_libs"}]

[tool.poetry.dependencies]
azure-storage-blob = "^12.21.0"
python = "^3.12"
slack-sdk = "^3.31.0"
tablib = "^3.6.1"

[tool.poetry.dev-dependencies]
bandit = "^1.7.8"
black = "^24.4.2"
coverage = "^7.6.0"
flake8 = "^7.1.0"
isort = "^5.13.2"
pydocstyle = "^6.3.0"
pylint = "^3.2.6"
pytest = "^8.3.1"
pytest-env = "^1.1.3"
toml = "^0.10.2"
Flake8-pyproject = "^1.2.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.bandit]
skips = ["B101"]
exclude_dirs = ["./tests/"]

[tool.black]
line-length = 120
include = '\.pyi?$'
exclude = '''
        /(
          \.git
        | \.tox
        | \.venv
        | env/
        | _build
        | build
        | dist
        )/
    '''

[tool.flake8]
ignore = ['E501', 'W503']

[tool.isort]
profile = "black"

[tool.pydocstyle]
convention = '''google'''
inherit = false
match = '''(?!__init__).*\.py'''
match-dir = '''(?!tests)[^\.].*'''

[tool.pylint.master]
init-hook = 'import os, sys; sys.path.append(os.path.dirname("aaron_common_libs/"));'

[tool.pylint.messages_control]
disable = '''
    line-too-long,
    singleton-comparison,
    too-many-lines,
    use-implicit-booleaness-not-comparison
    '''

[tool.pylint.miscellaneous]
notes = ''',
    TODO.md,
    '''

[tool.pylint.similarities]
min-similarity-lines = 10
# Required along with 'pylint: disable=duplicate-code' to ignore annoying messages.

[tool.pytest.ini_options]
pythonpath = [".", "aaron-common-libs/"]
env = ["LOG_LEVEL = DEBUG",
       "LOG_PATH = /tmp/",]